% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_test.R
\name{bootstrap_test}
\alias{bootstrap_test}
\title{Create bootstrap samples to calculate p-value}
\usage{
bootstrap_test(object, networks, attr = NULL, phicoef0 = NULL,
  phi0 = NULL, teststat, directed = FALSE, degree.spline = 3,
  interior.knot = 3, lambda.range = seq(-3, 3, by = 0.1),
  MCMC.burnin = 10000, MCMC.interval = 1000, NBoot = 1000, seed = 123)
}
\arguments{
\item{object}{A formula object of the form (network object) ~ <model terms>. Model terms take the same form of ERGM terms from R package 'ergm'.}

\item{networks}{A list of observed networks. It should have a list() object.}

\item{attr}{A list of vertex attributes. Default is NULL. (i.e. No attributes)}

\item{phicoef0}{Estimates of basis coefficients under H0}

\item{phi0}{Estimates of phi(t) under H0}

\item{teststat}{Test statistic calculated based on observed networks}

\item{directed}{TRUE for analyzing directed networks. FALSE for analyzing undirected networks.}

\item{degree.spline}{Degree of splines. Default is 3 (cubic splines).}

\item{interior.knot}{Number of interior knots to create splines. Default is 10.}

\item{lambda.range}{Range of lambda (Tuning parameter). Default is seq(-3, 3, by = 0.1).}

\item{MCMC.burnin}{MCMC burnin sample size. Default is 10000.}

\item{MCMC.interval}{Interval between selected networks. Default is 1000. The first simulated network is the (MCMC.burnin + MCMC.interval). Thereafter, every (MCMC.interval)th network will be sample.}

\item{NBoot}{Number of bootstrap samples. Default is 1000.}

\item{seed}{Seed number used to simulate bootstrap samples. Default is 123.}
}
\description{
Create bootstrap samples to calculate p-value
}

