% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_netstat.R
\name{calculate_netstat}
\alias{calculate_netstat}
\title{Calculate network statistics for temporal networks}
\usage{
calculate_netstat(networks, object, attr = NULL)
}
\arguments{
\item{networks}{A list of observed networks. It should have a list() object.}

\item{object}{A formula object of the form (network object) ~ <model terms>. Model terms take the same form of ERGM terms from R package 'ergm'.}

\item{attr}{Node-specific attributes. Default is NULL.}
}
\description{
Calculate network statistics for temporal networks
}
