% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_sectional_ergm.R
\name{cross.sectional_ergm}
\alias{cross.sectional_ergm}
\title{Calculate cross-sectional ERGM estimates}
\usage{
cross.sectional_ergm(object, networks, attr = NULL, directed = c(TRUE,
  FALSE), degree.spline = 3, interior.knot = 10, lambda.range = seq(-3, 3,
  by = 0.1))
}
\arguments{
\item{object}{A formula object of the form (network object) ~ <model terms>. Model terms take the same form of ERGM terms from R package 'ergm'.}

\item{networks}{A list of observed networks. It should have a list() object.}

\item{attr}{A list of vertex attributes. Default is NULL. (i.e. No attributes)}

\item{directed}{TRUE for analyzing directed networks. FALSE for analyzing undirected networks.}

\item{degree.spline}{Degree of splines. Default is 3 (cubic splines).}

\item{interior.knot}{Number of interior knots to create splines. Default is 10.}

\item{lambda.range}{Range of lambda (Tuning parameter). Default is seq(-3, 3, by = 0.1).}
}
\description{
Calculate cross-sectional ERGM estimates
}

