% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomOmega.R
\name{CustomOmega}
\alias{CustomOmega}
\title{Calculate Omega (Smoothness matrix)}
\usage{
CustomOmega(B, available.indx)
}
\arguments{
\item{B}{A set of basis functions (K x q matrix)}

\item{available.indx}{A list of observed time points}
}
\description{
Calculate Omega (Smoothness matrix)
}

