% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_adj_to_net.R
\name{from_adj_to_net}
\alias{from_adj_to_net}
\title{Convert a list of adjacency matrices into a list of network objects}
\usage{
from_adj_to_net(networks, attr = NULL, directed = FALSE)
}
\arguments{
\item{networks}{A list of adjacency matrices. It should have a list() object.}

\item{attr}{attr A list of vertex attributes. Default is NULL. (i.e. No attributes)}

\item{directed}{TRUE for analyzing directed networks. FALSE for analyzing undirected networks.}
}
\description{
Convert a list of adjacency matrices into a list of network objects
}
