% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof_vcergm.R
\name{gof_vcergm}
\alias{gof_vcergm}
\title{Goodness-of-fit diagnostics on VCERGM}
\usage{
gof_vcergm(object, object2 = NULL, networks, attr = NULL,
  directed = FALSE, netstat = NULL, degree.spline = 3,
  interior.knot = 10, nsim = 100, seed = 1234, log = FALSE)
}
\arguments{
\item{object}{A formula object of the form (network object) ~ <model terms>. Model terms take the same form of ERGM terms from R package 'ergm'.}

\item{object2}{A formula object of the form (network object) ~ <model terms> for GoF. By default, it is the same as object.}

\item{networks}{A list of observed networks. It should have a list() object.}

\item{attr}{A list of vertex attributes. Default is NULL. (i.e. No attributes)}

\item{directed}{TRUE for analyzing directed networks. FALSE for analyzing undirected networks.}

\item{netstat}{A vector of GoF network statistics names. Default is the name from ERGM estimates.}

\item{degree.spline}{Degree of splines. Default is 3 (cubic splines).}

\item{interior.knot}{Number of interior knots to create splines. Default is 10.}

\item{log}{Logarithm for observed and simulated network statistics. Default is FALSE.}
}
\description{
Goodness-of-fit diagnostics on VCERGM
}
