% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_network.R
\name{missing_network}
\alias{missing_network}
\title{Generate missing networks}
\usage{
missing_network(networks, num.missing = NULL, missing.time = NULL,
  seed = 1)
}
\arguments{
\item{networks}{A list of (fully) observed networks}

\item{num.missing}{Number of missing networks. Default is NULL.}

\item{missing.time}{A list of time points with missing networks. Default is NULL.}

\item{seed}{Seed number to generate missing.time when only num.missing is provided. Default seed number is 1.}
}
\description{
Generate missing networks
}
