% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mple_vcergm.R
\name{mple}
\alias{mple}
\title{Calculate the MPLE of a specified VCERGM}
\usage{
mple(object, networks, attr, directed, lag, B, degree.spline, lambda.range,
  constant, Tol)
}
\arguments{
\item{object}{A formula object of the form (network object) ~ <model terms>. Model terms take the same form of ERGM terms from R package 'ergm'.}

\item{networks}{A list of observed networks. It should have a list() object.}

\item{attr}{attr A list of vertex attributes. Default is NULL. (i.e. No attributes)}

\item{directed}{TRUE for analyzing directed networks. FALSE for analyzing undirected networks.}

\item{lag}{Lag in temporal networks. Default is 0 (i.e. VCERGM(0))}

\item{B}{A set of basis functions (K x q matrix)}

\item{degree.spline}{Degree of splines. Default is 3 (cubic splines).}

\item{lambda.range}{Range of lambda (Tuning parameter)}

\item{constant}{TRUE for constrained model of homogeneous VCERGM. FALSE for unconstrained model for heterogeneous VCERGM. Default is FALSE.}

\item{Tol}{Tolerance level used for calculating MPLE (IRLS iterations)}
}
\description{
Calculate the MPLE of a specified VCERGM
}
