% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{permutation_test}
\alias{permutation_test}
\title{Create permuted samples to calculate p-value}
\usage{
permutation_test(object, networks, attr = NULL, teststat, Delta,
  directed = FALSE, degree.spline = 3, interior.knot = 3,
  lambda.range = seq(-3, 3, by = 0.1), NPerm = 100, seed = 12345)
}
\arguments{
\item{object}{A formula object of the form (network object) ~ <model terms>. Model terms take the same form of ERGM terms from R package 'ergm'.}

\item{networks}{A list of observed networks. It should have a list() object.}

\item{attr}{A list of vertex attributes. Default is NULL. (i.e. No attributes)}

\item{teststat}{Test statistic calculated based on observed networks}

\item{directed}{TRUE for analyzing directed networks. FALSE for analyzing undirected networks.}

\item{degree.spline}{Degree of splines. Default is 3 (cubic splines).}

\item{interior.knot}{Number of interior knots to create splines. Default is 10.}

\item{lambda.range}{Range of lambda (Tuning parameter). Default is seq(-3, 3, by = 0.1).}

\item{NPerm}{Number of permuted samples. Default is 100.}

\item{seed}{Seed number used to simulate bootstrap samples. Default is 12345.}
}
\description{
Create permuted samples to calculate p-value
}
