% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_est2.R
\name{plotting2}
\alias{plotting2}
\title{Plot the (ERGM, VCERGM, TERGM) results}
\usage{
plotting2(ergmest, vcergmest, tergmest, interval = 10, xlab = "Time",
  label = NULL, timeseq = NULL, theme = NULL)
}
\arguments{
\item{ergmest}{ERGM estimate object}

\item{vcergmest}{VCERGM estimate object}

\item{tergmest}{TERGM estimate object}

\item{interval}{Interval size for time in x-axis. Default is 10.}

\item{xlab}{Name for x-axis. Default is "Time".}

\item{label}{Label(s) for plot title. Default is NULL (Use the label from VCERGM object).}

\item{timeseq}{User defined timeseq. Default is NULL (Time starts from 1).}

\item{theme}{ggplot theme. Default is NULL.}
}
\description{
Plot the (ERGM, VCERGM, TERGM) results
}
