% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_simulation.R
\name{plotting_simulation}
\alias{plotting_simulation}
\title{Plot the ERGM and VCERGM results when nsim > 1}
\usage{
plotting_simulation(object, true.phi, ergm.phi.hat, ergm.phi.hat.smooth,
  vcergm.phi.hat, interval = 10, label = NULL, quantile = FALSE)
}
\arguments{
\item{object}{A formula object of the form (network object) ~ <model terms>. Model terms take the same form of ERGM terms from R package 'ergm'.}

\item{true.phi}{TRUE phi(t) functions (number of network statistics x K vector)}

\item{ergm.phi.hat}{Estimated phi(t) from cross-sectional ERGMs. It should be a list of length nsim. For each simulation, the estimated phi.hat is a (number of network statistics x K) matrix.}

\item{ergm.phi.hat.smooth}{Estimated phi(t) from ad hoc 2-step procedure.}

\item{vcergm.phi.hat}{Estimated phi(t) from VCERGM.}
}
\description{
Plot the ERGM and VCERGM results when nsim > 1
}
