% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pspline.R
\name{Pspline}
\alias{Pspline}
\title{Penalized B-spline function}
\usage{
Pspline(y, H, lambda.range, B, available.indx, degree.spline)
}
\arguments{
\item{y}{Response matrix of the penalized logistic regression}

\item{H}{Design matrix of the penalized logistic regression}

\item{lambda.range}{Range of lambda (Tuning parameter)}

\item{B}{A set of basis functions (K x q matrix)}

\item{available.indx}{A list of observed time points}

\item{degree.spline}{Degree of splines. Default is 3 (cubic splines).}
}
\description{
Penalized B-spline function
}
