% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistic.R
\name{test_statistic}
\alias{test_statistic}
\title{Hypothesis test for heterogeneity}
\usage{
test_statistic(object, networks, attr = NULL, degree.spline = 3,
  interior.knot = 3, phicoef0 = NULL, phicoef1 = NULL, phi0 = NULL,
  phi1 = NULL, directed = c(TRUE, FALSE), Delta = NULL)
}
\arguments{
\item{object}{A formula object of the form (network object) ~ <model terms>. Model terms take the same form of ERGM terms from R package 'ergm'.}

\item{networks}{A list of observed networks. It should have a list() object.}

\item{attr}{A list of vertex attributes. Default is NULL. (i.e. No attributes)}

\item{degree.spline}{Degree of splines. Default is 3 (cubic splines).}

\item{interior.knot}{Number of interior knots to create splines. Default is 10.}

\item{phicoef0}{Estimates of basis coefficients under H0}

\item{phicoef1}{Estimates of basis coefficients under H1}

\item{phi0}{Estimates of phi(t) under H0}

\item{phi1}{Estimates of phi(t) under H1}

\item{directed}{TRUE for analyzing directed networks. FALSE for analyzing undirected networks.}
}
\description{
Hypothesis test for heterogeneity
}
