% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_est.R
\name{plotting}
\alias{plotting}
\title{Plot the ERGM and VCERGM results}
\usage{
plotting(ergmest, vcergmest, label = NULL)
}
\arguments{
\item{ergmest}{ERGM estimate object}

\item{vcergmest}{VCERGM estimate object}
}
\description{
Plot the ERGM and VCERGM results
}
